backend/requirements.txt
# Core Django
Django==5.2.10
djangorestframework==3.15.2

# CORS handling for React frontend
django-cors-headers==4.6.0

# Environment variables management
python-dotenv==1.0.1

# Groq LLM Integration
groq==0.13.0

# Data validation with Pydantic
pydantic==2.10.5
pydantic-settings==2.6.0

# Regex and validation utilities
email-validator==2.2.0
phonenumbers==8.13.54

# Sentiment analysis (lightweight)
textblob==0.18.0.post0

# JSON handling
orjson==3.10.14

# Date/Time utilities
python-dateutil==2.9.0.post0

# HTTP client for API calls
httpx==0.28.1

# Async support
asgiref==3.8.1

# Database (PostgreSQL for production - optional)
# psycopg2-binary==2.9.10

# Redis for session management (optional for production)
# redis==5.2.1
# django-redis==5.4.0

# Testing
pytest==8.3.4
pytest-django==4.9.0

# Code quality
black==24.10.0
flake8==7.1.1
